\name{compute_covariance}
\alias{compute_covariance}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
compute_covariance
}
\description{
Computes the covariance between two numeric vectors
}
\usage{
compute_covariance(x, y)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
  A numeric vector
}

  \item{y}{
  A numeric vector
}

}
\value{
Returns a real number representing the covariance
}
