\name{compute_variance}
\alias{compute_variance}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
compute_variance
}
\description{
Calculates the sample variance for a numeric vector
}
\usage{
compute_variance(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
  A numeric vector
}
}
\value{
Returns a real number representing the variance
}
